checkpoint: renames "TextToImage/config" to "TextToImage/Config"
