+++ src/features/TextToImage/Config/utilities/StaticConfig/index.ts
