+++ src/features/TextToImage/Config/utilities/emptyConfig.ts
